Moved utils out of pages directory
- Not routable
